requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.5
numpy>=1.24.0
pygame>=2.5.0
send2trash>=1.8.2
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
